// frontend-admin/src/api/generated/model/restoreVerificationRunResponseDtoStatus.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Kasse API
 * Registrierkasse API - RKSV uyumlu kasa sistemi
 * OpenAPI spec version: v1
 */

export type RestoreVerificationRunResponseDtoStatus = typeof RestoreVerificationRunResponseDtoStatus[keyof typeof RestoreVerificationRunResponseDtoStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const RestoreVerificationRunResponseDtoStatus = {
  NUMBER_0: 0,
  NUMBER_1: 1,
  NUMBER_2: 2,
  NUMBER_3: 3,
} as const;